requirements.txt
python-telegram-bot>=21.0
chromadb>=0.5.0
httpx>=0.27
PyPDF2>=3.0
python-docx>=1.0
chardet>=5.0
duckduckgo-search>=7.0
openpyxl>=3.1
beautifulsoup4>=4.12
